refactor: TicketDetail MVP 수정
