+++ app/src/main/java/woowacourse/movie/presentation/detail/TicketDetailActivity.kt
@@ -2,11 +2,12 @@ package woowacourse.movie.presentation.detail
 
 import android.os.Build
 import android.os.Bundle
+import android.util.Log
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
+import woowacourse.movie.presentation.model.TicketModel
 import woowacourse.movie.presentation.reservation.MovieReservationPresenter
-import woowacourse.movie.presentation.reservation.model.TicketModel
 import woowacourse.movie.presentation.utils.toCustomString
 
 class TicketDetailActivity : AppCompatActivity(), TicketDetailContract.View {
@@ -14,6 +15,9 @@ class TicketDetailActivity : AppCompatActivity(), TicketDetailContract.View {
     private lateinit var ticketScreeningDate: TextView
     private lateinit var ticketPrice: TextView
     private lateinit var ticketCount: TextView
+    private lateinit var ticketTime: TextView
+    private lateinit var ticketSeat: TextView
+
     private val presenter: TicketDetailPresenter by lazy {
         TicketDetailPresenter(
             view = this@TicketDetailActivity,
@@ -33,6 +37,8 @@ class TicketDetailActivity : AppCompatActivity(), TicketDetailContract.View {
         ticketScreeningDate = findViewById(R.id.ticket_screening_date)
         ticketPrice = findViewById(R.id.ticket_price)
         ticketCount = findViewById(R.id.ticket_number_of_people)
+        ticketTime = findViewById(R.id.ticket_screening_time)
+        ticketSeat = findViewById(R.id.ticket_seat)
     }
 
     private fun getReservationTicket(): TicketModel {
@@ -51,6 +57,12 @@ class TicketDetailActivity : AppCompatActivity(), TicketDetailContract.View {
         ticketTitle.text = ticketModel.title
         ticketScreeningDate.text = ticketModel.screeningDate.toCustomString()
         ticketPrice.text = getString(R.string.ticket_price, ticketModel.price)
-        ticketCount.text = getString(R.string.ticket_count,ticketModel.count)
+        ticketCount.text = getString(R.string.ticket_count, ticketModel.count)
+        ticketTime.text = ticketModel.screeningTime.toCustomString()
+        ticketSeat.text = ticketModel.seats.joinToString(SEAT_SEPARATOR)
+    }
+
+    companion object {
+        private const val SEAT_SEPARATOR = ", "
     }
 }
