+++ app/src/main/java/woowacourse/movie/presentation/detail/TicketDetailContract.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.presentation.detail
 
-import woowacourse.movie.presentation.reservation.model.TicketModel
+import woowacourse.movie.presentation.model.TicketModel
 
 interface TicketDetailContract {
     interface View {
