+++ app/src/main/java/woowacourse/movie/presentation/detail/TicketDetailPresenter.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.presentation.detail
 
-import woowacourse.movie.presentation.reservation.model.TicketModel
+import woowacourse.movie.presentation.model.TicketModel
 
 class TicketDetailPresenter(
     private val view: TicketDetailContract.View,
